Move JsonStyleSelector to its own file.
