+++ Sandra.UI/AppTemplate/JsonSyntaxDescriptor.cs
@@ -25,7 +25,6 @@ using Eutherion.Win.Storage;
 using ScintillaNET;
 using System;
 using System.Collections.Generic;
-using System.Drawing;
 
 namespace Eutherion.Win.AppTemplate
 {
@@ -95,56 +94,5 @@ namespace Eutherion.Win.AppTemplate
 
         public override string GetErrorMessage(JsonErrorInfo error)
             => error.Message(Session.Current.CurrentLocalizer);
-    }
-
-    /// <summary>
-    /// A style selector for json syntax highlighting.
-    /// </summary>
-    public class JsonStyleSelector : JsonSymbolVisitor<SyntaxEditor<SettingSyntaxTree, JsonSymbol, JsonErrorInfo>, Style>
-    {
-        private const int commentStyleIndex = 8;
-        private const int valueStyleIndex = 9;
-        private const int stringStyleIndex = 10;
-
-        private static readonly Color commentForeColor = Color.FromArgb(128, 220, 220);
-        private static readonly Font commentFont = new Font("Consolas", 10, FontStyle.Italic);
-
-        private static readonly Color valueForeColor = Color.FromArgb(255, 255, 60);
-        private static readonly Font valueFont = new Font("Consolas", 10, FontStyle.Bold);
-
-        private static readonly Color stringForeColor = Color.FromArgb(255, 192, 144);
-
-        public static readonly JsonStyleSelector Instance = new JsonStyleSelector();
-
-        public static void InitializeStyles(SyntaxEditor<SettingSyntaxTree, JsonSymbol, JsonErrorInfo> syntaxEditor)
-        {
-            syntaxEditor.Styles[commentStyleIndex].ForeColor = commentForeColor;
-            commentFont.CopyTo(syntaxEditor.Styles[commentStyleIndex]);
-
-            syntaxEditor.Styles[valueStyleIndex].ForeColor = valueForeColor;
-            valueFont.CopyTo(syntaxEditor.Styles[valueStyleIndex]);
-
-            syntaxEditor.Styles[stringStyleIndex].ForeColor = stringForeColor;
-        }
-
-        private JsonStyleSelector() { }
-
-        public override Style DefaultVisit(JsonSymbol symbol, SyntaxEditor<SettingSyntaxTree, JsonSymbol, JsonErrorInfo> syntaxEditor)
-            => syntaxEditor.DefaultStyle;
-
-        public override Style VisitComment(JsonComment symbol, SyntaxEditor<SettingSyntaxTree, JsonSymbol, JsonErrorInfo> syntaxEditor)
-            => syntaxEditor.Styles[commentStyleIndex];
-
-        public override Style VisitErrorString(JsonErrorString symbol, SyntaxEditor<SettingSyntaxTree, JsonSymbol, JsonErrorInfo> syntaxEditor)
-            => syntaxEditor.Styles[stringStyleIndex];
-
-        public override Style VisitString(JsonString symbol, SyntaxEditor<SettingSyntaxTree, JsonSymbol, JsonErrorInfo> syntaxEditor)
-            => syntaxEditor.Styles[stringStyleIndex];
-
-        public override Style VisitUnterminatedMultiLineComment(JsonUnterminatedMultiLineComment symbol, SyntaxEditor<SettingSyntaxTree, JsonSymbol, JsonErrorInfo> syntaxEditor)
-            => syntaxEditor.Styles[commentStyleIndex];
-
-        public override Style VisitValue(JsonValue symbol, SyntaxEditor<SettingSyntaxTree, JsonSymbol, JsonErrorInfo> syntaxEditor)
-            => syntaxEditor.Styles[valueStyleIndex];
     }
 }
