+++ Sandra.UI/Sandra.UI.csproj
@@ -57,6 +57,7 @@
     <Compile Include="AppTemplate\FileLocalizer.cs" />
     <Compile Include="AppTemplate\FormStateAutoSaver.cs" />
     <Compile Include="AppTemplate\JsonErrorInfoExtensions.cs" />
+    <Compile Include="AppTemplate\JsonStyleSelector.cs" />
     <Compile Include="AppTemplate\JsonSyntaxDescriptor.cs" />
     <Compile Include="AppTemplate\ListBoxEx.cs">
       <SubType>Component</SubType>
